migrate pkgs from compat to chain
