+++ chain/digibyte/address.go
@@ -1,0 +1,7 @@
+package digibyte
+
+import "github.com/renproject/multichain/chain/bitcoin"
+
+type AddressEncoder = bitcoin.AddressEncoder
+type AddressDecoder = bitcoin.AddressDecoder
+type AddressEncodeDecoder = bitcoin.AddressEncodeDecoder
