+++ chain/digibyte/address_test.go
@@ -1,9 +1,9 @@
-package digibyte_test
+package bitcoin_test
 
 import (
+	"github.com/renproject/multichain/api/address"
 	"github.com/renproject/multichain/chain/bitcoin"
 	"github.com/renproject/multichain/chain/digibyte"
-	"github.com/renproject/pack"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -12,7 +12,7 @@ import (
 var _ = Describe("DigiByte", func() {
 	Context("when decoding an address", func() {
 		It("should work without errors", func() {
-			_, err := bitcoin.NewAddressDecoder(&digibyte.MainNetParams).DecodeAddress(pack.NewString("DBLsEv4FdFPGrMWzcagDQvoKgUL2CikhMf"))
+			_, err := bitcoin.NewAddressDecoder(&digibyte.MainNetParams).DecodeAddress(address.Address("DBLsEv4FdFPGrMWzcagDQvoKgUL2CikhMf"))
 			Expect(err).NotTo(HaveOccurred())
 		})
 	})
