+++ chain/digibyte/gas.go
@@ -1,0 +1,5 @@
+package digibyte
+
+import "github.com/renproject/multichain/chain/bitcoin"
+
+type GasEstimator = bitcoin.GasEstimator
