+++ chain/digibyte/gas_test.go
@@ -1,0 +1,1 @@
+package digibyte_test
