+++ chain/digibyte/utxo.go
@@ -1,0 +1,9 @@
+package digibyte
+
+import "github.com/renproject/multichain/chain/bitcoin"
+
+type Tx = bitcoin.Tx
+
+type TxBuilder = bitcoin.TxBuilder
+
+type Client = bitcoin.Client
