+++ chain/dogecoin/gas.go
@@ -1,0 +1,5 @@
+package dogecoin
+
+import "github.com/renproject/multichain/chain/bitcoin"
+
+type GasEstimator = bitcoin.GasEstimator
