+++ chain/dogecoin/utxo.go
@@ -1,0 +1,9 @@
+package dogecoin
+
+import "github.com/renproject/multichain/chain/bitcoin"
+
+type Tx = bitcoin.Tx
+
+type TxBuilder = bitcoin.TxBuilder
+
+type Client = bitcoin.Client
