+++ chain/ethereum/account_test.go
@@ -1,0 +1,1 @@
+package ethereum_test
