+++ chain/ethereum/address_test.go
@@ -1,11 +1,11 @@
-package ethereumcompat_test
+package ethereum_test
 
 import (
 	"encoding/hex"
 	"encoding/json"
 	"testing/quick"
 
-	"github.com/renproject/multichain/compat/ethereumcompat"
+	"github.com/renproject/multichain/chain/ethereum"
 	"github.com/renproject/surge"
 
 	. "github.com/onsi/ginkgo"
@@ -16,13 +16,13 @@ var _ = Describe("Address", func() {
 	Context("when unmarshaling and unmarshaling", func() {
 		It("should equal itself", func() {
 			f := func(x [20]byte) bool {
-				addr := ethereumcompat.Address(x)
+				addr := ethereum.Address(x)
 				Expect(addr.SizeHint()).To(Equal(20))
 
 				bytes, err := surge.ToBinary(addr)
 				Expect(err).ToNot(HaveOccurred())
 
-				var newAddr ethereumcompat.Address
+				var newAddr ethereum.Address
 				err = surge.FromBinary(&newAddr, bytes)
 				Expect(err).ToNot(HaveOccurred())
 
@@ -38,12 +38,12 @@ var _ = Describe("Address", func() {
 	Context("when unmarshaling and unmarshaling to/from JSON", func() {
 		It("should equal itself", func() {
 			f := func(x [20]byte) bool {
-				addr := ethereumcompat.Address(x)
+				addr := ethereum.Address(x)
 
 				bytes, err := json.Marshal(addr)
 				Expect(err).ToNot(HaveOccurred())
 
-				var newAddr ethereumcompat.Address
+				var newAddr ethereum.Address
 				err = json.Unmarshal(bytes, &newAddr)
 				Expect(err).ToNot(HaveOccurred())
 
@@ -61,7 +61,7 @@ var _ = Describe("Address", func() {
 					bytes, err := json.Marshal(string(x[:]))
 					Expect(err).ToNot(HaveOccurred())
 
-					var newAddr ethereumcompat.Address
+					var newAddr ethereum.Address
 					err = json.Unmarshal(bytes, &newAddr)
 					Expect(err).To(HaveOccurred())
 					return true
@@ -79,7 +79,7 @@ var _ = Describe("Address", func() {
 					bytes, err := json.Marshal(addr)
 					Expect(err).ToNot(HaveOccurred())
 
-					var newAddr ethereumcompat.Address
+					var newAddr ethereum.Address
 					err = json.Unmarshal(bytes, &newAddr)
 					Expect(err).To(HaveOccurred())
 					return true
@@ -94,7 +94,7 @@ var _ = Describe("Address", func() {
 	Context("when unmarshalling random data", func() {
 		It("should not panic", func() {
 			f := func(x []byte) bool {
-				var addr ethereumcompat.Address
+				var addr ethereum.Address
 				Expect(func() { addr.Unmarshal(x, surge.MaxBytes) }).ToNot(Panic())
 				Expect(func() { json.Unmarshal(x, &addr) }).ToNot(Panic())
 				return true
