+++ chain/substrate/address_test.go
@@ -1,0 +1,1 @@
+package substrate_test
