--- compat/bitcoincompat/bitcoincompat_suite_test.go
@@ -1,13 +1,0 @@
-package bitcoincompat_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestBitcoinCompat(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Bitcoin Compat Suite")
-}
