--- compat/ethereumcompat/client_test.go
@@ -1,1 +1,0 @@
-package ethereumcompat_test
