--- compat/ethereumcompat/ethereumcompat.go
@@ -1,1 +1,0 @@
-package ethereumcompat
