--- compat/ethereumcompat/ethereumcompat_suite_test.go
@@ -1,13 +1,0 @@
-package ethereumcompat_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestEthereumCompat(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Ethereum Compat Suite")
-}
