--- compat/substratecompat/client_test.go
@@ -1,1 +1,0 @@
-package substratecompat_test
