--- compat/substratecompat/substratecompat.go
@@ -1,1 +1,0 @@
-package substratecompat
